+++ lib/pavilion/clean.py
@@ -6,9 +6,9 @@ from typing import List
 
 from pavilion import dir_db
 from pavilion import lockfile
-from pavilion import test_run
 from pavilion import utils
 from pavilion.builder import TestBuilder
+from pavilion.test_run import test_run_attr_transform
 
 
 def delete_tests(id_dir: Path, filter_func, verbose: bool = False):
@@ -17,7 +17,7 @@ def delete_tests(id_dir: Path, filter_func, verbose: bool = False):
     if filter_func is None:
         filter_func = dir_db.default_filter
     return dir_db.delete(id_dir, filter_func,
-                         transform=test_run.test_run_attr_transform,
+                         transform=test_run_attr_transform,
                          verbose=verbose)
