+++ lib/pavilion/series/series.py
@@ -2,13 +2,12 @@
 also tracks the tests that have run under it."""
 
 import json
-import logging
 import os
 import subprocess
 import time
-from collections import defaultdict
+from collections import defaultdict, UserDict
 from pathlib import Path
-from typing import List, Dict, Set, Tuple, Union, TextIO
+from typing import List, Dict, Set, Union, TextIO
 
 from pavilion import dir_db
 from pavilion import output
@@ -19,10 +18,54 @@ from pavilion.lockfile import LockFile
 from pavilion.output import fprint
 from pavilion.series_config import SeriesConfigLoader
 from pavilion.status_file import STATES
-from pavilion.test_run import TestRun, TestRunError
+from pavilion.test_run import TestRun, ID_Pair
 from yaml_config import YAMLError, RequiredError
 from .errors import TestSeriesError, TestSeriesWarning
 from .test_set import TestSet, TestSetError
+
+
+class LazyTestRunDict(UserDict):
+    """A lazily evaluated dictionary of tests."""
+
+    def __init__(self, pav_cfg):
+        """Initialize the lazy TestRun dict."""
+
+        self._pav_cfg = pav_cfg
+
+        super().__init__()
+
+    def add_key(self, id_pair: ID_Pair):
+        """Add an ID_Pair key, but don't actually load the test."""
+
+        self.data[id_pair] = None
+
+    def __getitem__(self, id_pair: ID_Pair) -> TestRun:
+        """When the item exists as a key but not a test object, load the test object."""
+
+        if id_pair in self.data and self.data[id_pair] is None:
+            working_dir, test_id = id_pair
+            self.data[id_pair] = TestRun.load(self._pav_cfg, working_dir, test_id)
+
+        return super().__getitem__(id_pair)
+
+    def find_tests(self, series_path: Path):
+        """Find all the tests for the series and add their keys."""
+
+        for path in dir_db.select(series_path, use_index=False).paths:
+            if not path.is_symlink():
+                continue
+
+            try:
+                test_id = int(path.name)
+            except ValueError:
+                continue
+
+            try:
+                working_dir = path.resolve().parents[1]
+            except FileNotFoundError:
+                continue
+
+            self.add_key(ID_Pair((working_dir, test_id)))
 
 
 class TestSeries:
@@ -51,7 +94,8 @@ class TestSeries:
         """
 
         self.pav_cfg = pav_cfg
-        self.tests = {}  # type: Dict[Tuple[Path, int], TestRun]
+        self.tests = LazyTestRunDict(pav_cfg)
+
         self.config = config or SeriesConfigLoader().load_empty()
 
         series_path = self.pav_cfg.working_dir/'series'
@@ -84,8 +128,6 @@ class TestSeries:
         else:
             self._id = _id
             self.path = dir_db.make_id_path(series_path, self._id)
-
-        self._logger = logging.getLogger(self.LOGGER_FMT.format(self._id))
 
     def run_background(self):
         """Run pav _series in background using subprocess module."""
@@ -182,34 +224,6 @@ differentiate it from test ids."""
             raise TestSeriesError("No such series found: '{}' at '{}'"
                                   .format(series_id, series_path))
 
-        logger = logging.getLogger(cls.LOGGER_FMT.format(series_id))
-
-        tests = []
-        for path in dir_db.select(series_path, use_index=False).paths:
-            try:
-                test_id = int(path.name)
-            except ValueError:
-                series_info_files = [cls.OUT_FN,
-                                     cls.PGID_FN,
-                                     cls.CONFIG_FN,
-                                     cls.DEPENDENCY_FN]
-                if path.name not in series_info_files:
-                    logger.info("Bad test id in series from dir '%s'", path)
-                continue
-
-            try:
-                working_dir = path.resolve().parents[1]
-            except FileNotFoundError as err:
-                logger.info("Bad test id in series %s: %s", sid, err.args[0])
-                continue
-
-            try:
-                test = TestRun.load(pav_cfg, working_dir, test_id)
-                tests.append(test)
-            except TestRunError as err:
-                logger.info("Error loading test %s: %s",
-                            test_id, err.args[0])
-
         loader = SeriesConfigLoader()
         try:
             with (series_path/cls.CONFIG_FN).open() as config_file:
@@ -223,7 +237,9 @@ differentiate it from test ids."""
             raise TestSeriesError("Could not load config file for test series '{}': {}"
                                   .format(sid, err.args[0]))
 
-        return cls(pav_cfg, _id=series_id, config=config)
+        series = cls(pav_cfg, _id=series_id, config=config)
+        series.tests.find_tests(series.path)
+        return series
 
     def _create_test_sets(self):
         """Create test sets from the config, and set up their dependency
@@ -565,7 +581,7 @@ differentiate it from test ids."""
         # attempt to make symlink
         link_path = dir_db.make_id_path(self.path, test.id)
 
-        self.tests[(test.working_dir, test.id)] = test
+        self.tests[test.id_pair] = test
 
         if not link_path.exists():
             try:
