+++ test/tests/general_tests.py
@@ -7,7 +7,7 @@ from pathlib import Path
 from typing import List
 
 import yc_yaml as yaml
-from pavilion import test_run
+from pavilion.test_run import TestRun
 from pavilion import utils
 from pavilion.unittest import PavTestCase
 
@@ -124,7 +124,7 @@ class GeneralTests(PavTestCase):
             build_dst = dst_path/build_dst
             (dst_path/'build_dir').rename(build_dst)
 
-            test = test_run.TestRun.load_from_raw_id(self.pav_cfg, run_id)
+            test = TestRun.load_from_raw_id(self.pav_cfg, run_id)
             self.assertTrue(test.results)
             self.assertTrue(test.complete)
