+++ test/tests/raw_sched_tests.py
@@ -1,10 +1,6 @@
 from pavilion import plugins
 from pavilion import schedulers
-from pavilion.test_config import variables
-from pavilion.test_run import TestRun
 from pavilion.unittest import PavTestCase
-from pavilion.test_config import VariableSetManager
-import re
 
 
 class RawSchedTests(PavTestCase):
