+++ test/tests/test_run_tests.py
@@ -2,7 +2,8 @@ import io
 
 from pavilion import plugins
 from pavilion.test_config import VariableSetManager, TestConfigResolver
-from pavilion.test_run import TestRunError, TestRun
+from pavilion.test_run import TestRun
+from pavilion.exceptions import TestRunError
 from pavilion.unittest import PavTestCase
